SSSP root set to Long

+++ src/main/scala/graph/AlgRunner.scala
@@ -18,8 +18,8 @@ object AlgRunner {
     algOpt.split(":").toList match {
       case "bfs" :: root :: Nil => engine.run(new BFS(root.toLong))
       case "bfs" :: "u" :: root :: Nil => engine.run(new BFS_U(root.toLong))
-      case "sssp" :: root :: Nil => engine_w.run(new SSSP(root.toInt))
-      case "sssp" :: "u" :: root :: Nil => engine_w.run(new SSSP_U(root.toInt))
+      case "sssp" :: root :: Nil => engine_w.run(new SSSP(root.toLong))
+      case "sssp" :: "u" :: root :: Nil => engine_w.run(new SSSP_U(root.toLong))
       case "cc" :: Nil => engine.run(new CC)
       case "kcore" :: Nil => engine.run(new KCore)
       case "pagerank" :: nLoop :: Nil => engine.run(new PageRank(nLoop.toInt))
